refactor: related drop notification component
